api/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.10.0
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
distlib==0.4.0
elasticsearch==8.9.0
fastapi==0.116.1
filelock==3.19.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
identify==2.6.13
idna==3.10
isort==6.0.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pre_commit==4.3.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
